+++ src/ContainerBuilder.php
@@ -1,4 +1,5 @@
 <?php
+declare(strict_types=1);
 /**
  *
  * This file is part of Aura for PHP.
@@ -40,7 +41,7 @@ class ContainerBuilder
      * @return Container
      *
      */
-    public function newInstance($autoResolve = false)
+    public function newInstance(bool $autoResolve = false): Container
     {
         $resolver = $this->newResolver($autoResolve);
         return new Container(new InjectionFactory($resolver));
@@ -55,7 +56,7 @@ class ContainerBuilder
      * @return Resolver
      *
      */
-    protected function newResolver($autoResolve = false)
+    protected function newResolver(bool $autoResolve = false): Resolver
     {
         if ($autoResolve) {
             return new AutoResolver(new Reflector());
@@ -82,8 +83,8 @@ class ContainerBuilder
      */
     public function newConfiguredInstance(
         array $configClasses = [],
-        $autoResolve = false
-    ) {
+        bool $autoResolve = false
+    ): Container {
         $di = $this->newInstance($autoResolve);
         $collection = $this->newConfigCollection($configClasses);
 
@@ -105,7 +106,7 @@ class ContainerBuilder
      *
      * @return ConfigCollection
      */
-    protected function newConfigCollection(array $configClasses = [])
+    protected function newConfigCollection(array $configClasses = []): ConfigCollection
     {
         return new ConfigCollection($configClasses);
     }