+++ src/Injection/Factory.php
@@ -1,4 +1,5 @@
 <?php
+declare(strict_types=1);
 /**
  *
  * This file is part of Aura for PHP.
@@ -72,7 +73,7 @@ class Factory
      */
     public function __construct(
         Resolver $resolver,
-        $class,
+        string $class,
         array $params = [],
         array $setters = []
     ) {
@@ -95,7 +96,7 @@ class Factory
      * @return object
      *
      */
-    public function __invoke()
+    public function __invoke(): object
     {
         $params = array_merge($this->params, func_get_args());
         $resolve = $this->resolver->resolve(