+++ tests/Resolver/ResolverTest.php
@@ -2,8 +2,9 @@
 namespace Aura\Di\Resolver;
 
 use Aura\Di\Injection\Lazy;
+use PHPUnit\Framework\TestCase;
 
-class ResolverTest extends \PHPUnit_Framework_TestCase
+class ResolverTest extends TestCase
 {
     protected $resolver;
 
@@ -16,7 +17,7 @@ class ResolverTest extends \PHPUnit_Framework_TestCase
     public function testReadsConstructorDefaults()
     {
         $expect = ['foo' => 'bar'];
-        list($actual_params, $actual_setter) = $this->resolver->getUnified('Aura\Di\Fake\FakeParentClass');
+        [$actual_params, ] = $this->resolver->getUnified('Aura\Di\Fake\FakeParentClass');
         $this->assertSame($expect, $actual_params);
     }
 
@@ -34,7 +35,7 @@ class ResolverTest extends \PHPUnit_Framework_TestCase
             'zim' => null,
         ];
 
-        list($actual_params, $actual_setter) = $this->resolver->getUnified('Aura\Di\Fake\FakeChildClass');
+        [$actual_params, ] = $this->resolver->getUnified('Aura\Di\Fake\FakeChildClass');
         $this->assertSame($expect, $actual_params);
     }
 
@@ -43,7 +44,7 @@ class ResolverTest extends \PHPUnit_Framework_TestCase
         $this->resolver->params['Aura\Di\Fake\FakeParentClass'] = ['foo' => 'zim'];
 
         $expect = ['foo' => 'zim'];
-        list($actual_params, $actual_setter) = $this->resolver->getUnified('Aura\Di\Fake\FakeParentClass');
+        [$actual_params, ] = $this->resolver->getUnified('Aura\Di\Fake\FakeParentClass');
         $this->assertSame($expect, $actual_params);
     }
 
@@ -56,7 +57,7 @@ class ResolverTest extends \PHPUnit_Framework_TestCase
             'zim' => null,
         ];
 
-        list($actual_params, $actual_setter) = $this->resolver->getUnified('Aura\Di\Fake\FakeChildClass');
+        [$actual_params, ] = $this->resolver->getUnified('Aura\Di\Fake\FakeChildClass');
         $this->assertSame($expect, $actual_params);
 
         // for test coverage of the mock class
@@ -67,7 +68,7 @@ class ResolverTest extends \PHPUnit_Framework_TestCase
     {
         $this->resolver->setters['Aura\Di\Fake\FakeParentClass']['setFake'] = 'fake1';
 
-        list($actual_config, $actual_setter) = $this->resolver->getUnified('Aura\Di\Fake\FakeChildClass');
+        [, $actual_setter] = $this->resolver->getUnified('Aura\Di\Fake\FakeChildClass');
         $expect = ['setFake' => 'fake1'];
         $this->assertSame($expect, $actual_setter);
 
@@ -78,7 +79,7 @@ class ResolverTest extends \PHPUnit_Framework_TestCase
         $this->resolver->setters['Aura\Di\Fake\FakeParentClass']['setFake'] = 'fake1';
         $this->resolver->setters['Aura\Di\Fake\FakeChildClass']['setFake'] = 'fake2';
 
-        list($actual_config, $actual_setter) = $this->resolver->getUnified('Aura\Di\Fake\FakeChildClass');
+        [, $actual_setter] = $this->resolver->getUnified('Aura\Di\Fake\FakeChildClass');
         $expect = ['setFake' => 'fake2'];
         $this->assertSame($expect, $actual_setter);
     }
@@ -87,7 +88,7 @@ class ResolverTest extends \PHPUnit_Framework_TestCase
     {
         $this->resolver->setters['Aura\Di\Fake\FakeTrait']['setFake'] = 'fake1';
 
-        list($actual_config, $actual_setter) = $this->resolver->getUnified('Aura\Di\Fake\FakeClassWithTrait');
+        [, $actual_setter] = $this->resolver->getUnified('Aura\Di\Fake\FakeClassWithTrait');
         $expect = ['setFake' => 'fake1'];
         $this->assertSame($expect, $actual_setter);
     }
@@ -96,7 +97,7 @@ class ResolverTest extends \PHPUnit_Framework_TestCase
     {
         $this->resolver->setters['Aura\Di\Fake\FakeChildTrait']['setChildFake'] = 'fake1';
 
-        list($actual_config, $actual_setter) = $this->resolver->getUnified('Aura\Di\Fake\FakeClassWithTrait');
+        [, $actual_setter] = $this->resolver->getUnified('Aura\Di\Fake\FakeClassWithTrait');
         $expect = ['setChildFake' => 'fake1'];
         $this->assertSame($expect, $actual_setter);
     }
@@ -105,9 +106,7 @@ class ResolverTest extends \PHPUnit_Framework_TestCase
     {
         $this->resolver->setters['Aura\Di\Fake\FakeGrandchildTrait']['setGrandchildFake'] = 'fake1';
 
-        list($actual_config, $actual_setter) = $this->resolver->getUnified(
-            'Aura\Di\Fake\FakeClassWithTrait'
-        );
+        [, $actual_setter] = $this->resolver->getUnified('Aura\Di\Fake\FakeClassWithTrait');
         $expect = ['setGrandchildFake' => 'fake1'];
         $this->assertSame($expect, $actual_setter);
     }
@@ -115,9 +114,7 @@ class ResolverTest extends \PHPUnit_Framework_TestCase
     public function testHonorsParentClassTraits()
     {
         $this->resolver->setters['Aura\Di\Fake\FakeGrandchildTrait']['setGrandchildFake'] = 'fake1';
-        list($actual_config, $actual_setter) = $this->resolver->getUnified(
-            'Aura\Di\Fake\FakeClassWithParentTrait'
-        );
+        [, $actual_setter] = $this->resolver->getUnified('Aura\Di\Fake\FakeClassWithParentTrait');
         $expect = ['setGrandchildFake' => 'fake1'];
         $this->assertSame($expect, $actual_setter);
     }
@@ -129,14 +126,14 @@ class ResolverTest extends \PHPUnit_Framework_TestCase
         $this->resolver->setters['Aura\Di\Fake\FakeClassWithTrait']['setFake'] = 'fake3';
         $this->resolver->setters['Aura\Di\Fake\FakeClassWithTrait']['setChildFake'] = 'fake4';
 
-        list($actual_config, $actual_setter) = $this->resolver->getUnified('Aura\Di\Fake\FakeClassWithTrait');
+        [, $actual_setter] = $this->resolver->getUnified('Aura\Di\Fake\FakeClassWithTrait');
         $expect = ['setChildFake' => 'fake4', 'setFake' => 'fake3'];
         $this->assertSame($expect, $actual_setter);
     }
 
     public function testReflectionOnMissingClass()
     {
-        $this->setExpectedException('ReflectionException');
+        $this->expectException('ReflectionException');
         $this->resolver->resolve('NoSuchClass');
     }
 
@@ -151,16 +148,14 @@ class ResolverTest extends \PHPUnit_Framework_TestCase
 
     public function testMissingParam()
     {
-        $this->setExpectedException(
-            'Aura\Di\Exception\MissingParam',
-            'Aura\Di\Fake\FakeResolveClass::$fake'
-        );
+        $this->expectException('Aura\Di\Exception\MissingParam');
+        $this->expectExceptionMessage('Aura\Di\Fake\FakeResolveClass::$fake');
         $this->resolver->resolve('Aura\Di\Fake\FakeResolveClass');
     }
 
     public function testUnresolvedParamAfterMergeParams()
     {
-        $this->setExpectedException('Aura\Di\Exception\MissingParam');
+        $this->expectException('Aura\Di\Exception\MissingParam');
         $this->resolver->resolve('Aura\Di\Fake\FakeParamsClass', [
             'noSuchParam' => 'foo'
         ]);