+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG
 
+## 4.0.0
+
+- Upgrade to PHP 7.2+ only, strict types enabled
+- Dropped HHVM support
+- Signatures of methods in ContainerConfigInterface changed.
+- LazyRequire and LazyInclude do not accept another lazy
+- Removed container interop dependency
+
 ## 3.4.0
 
 - (CHG) LazyArray now extends ArrayObject. PR #151.